requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.2.0
blinker==1.7.0
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.109.1
Flask==3.0.2
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pydantic==2.6.0
pydantic_core==2.16.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.27.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
